Fix Basic/Exercise-12

+++ src/main/java/basic/Exercise12.java
@@ -2,9 +2,9 @@ package basic;
 
 public class Exercise12 {
 
-    public float findSquareRoot(float num) {
-        float guess = num;
-        while (Math.abs(guess * guess - num) / num >= 0.0001f)
+    public double findSquareRoot(double num) {
+        double guess = num;
+        while (Math.abs(guess * guess - num) / num >= 0.0001)
             guess = num / (2 * guess) + guess / 2;
         return guess;
     }
